requirements.txt
python-nmap>=0.6.1
requests>=2.10.0
tabulate>=0.7.5
